Fix handling of delayed exit option (Closes: #849192, LP: #1656785)
The option for delayed exit (-q#) existed before but since upstream
rewrote the readwrite call the patch adding it had to be adapted.
However the current implementation causes the loop to exit too early.
Move the exit handling to those places which do exit the loop in the
original code. While on it also add some code to close the net_fd on
delayed exit.